chore: move AR exception interface into Exception namespace
