+++ src/Exception/ActiveRecordException.php
@@ -2,6 +2,6 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ActiveRecord\Contract;
+namespace Cycle\ActiveRecord\Exception;
 
 interface ActiveRecordException extends \Throwable {}
