+++ src/Exception/ConfigurationException.php
@@ -4,6 +4,4 @@ declare(strict_types=1);
 
 namespace Cycle\ActiveRecord\Exception;
 
-use Cycle\ActiveRecord\Contract\ActiveRecordException;
-
 class ConfigurationException extends \RuntimeException implements ActiveRecordException {}
